requirements.txt
# LangChain and LangGraph
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
langchain-google-genai>=2.0.0

# LLM Providers
openai>=1.50.0
anthropic>=0.39.0
google-generativeai>=0.8.0

# Search Tools
tavily-python>=0.5.0
arxiv>=2.1.0

# UI
rich>=13.0.0

# HTTP and Async
httpx>=0.27.0
requests>=2.32.0

# Data Validation and Configuration
pydantic>=2.9.0
python-dotenv>=1.0.0
jinja2>=3.1.0

# Utilities
typing-extensions>=4.12.0
